backend/mcp/requirements.txt
fastapi
uvicorn
langchain
python-dotenv
langchain-openrouter
langchain-openai